Fix minor side effect behavior with major refactorization of CookieBanner.tsx
The GA script produced a side effect that set the banner checks intrusively (because of cookie change events). The component is now refactored into small, better-designed components, and easy to test. Now, the CookieAction component is lower-level, while the CookieBanner parent takes care of its default state and the banner's conditional rendering handling.