Avoid prompting CookieBanner if CookieCustomization pane was showed
The banner prompted redundantly after the customization pane was open and then closed the first time when there's no consent stored yet. Either if the user takes action or not in the pane, it's undesired/incorrect to show the banner again when closing the pane. If the user chose nothing from the pane, the banner will re-prompt until the next page-reload. If the user gave consent (the first time) from the pane, the request is processed asynchronously so the banner must not show even if the consent is applied locally later.